update PKG_CONFIG_PATH, PATH variables

+++ Bash/Installer Scripts/GitHub Projects/build-libboost.sh
@@ -26,23 +26,19 @@ printf "%s\n%s\n\n" \
 
 # Create output directory
 
-if [ -d "$cwd" ]; then
-    sudo rm -fr "$cwd"
-fi
+[[ -d "$cwd" ]] && sudo rm -fr "$cwd"
 mkdir -p "$cwd"
 
 # Set the c + cpp compilers
-
-export CC=gcc CXX=g++
-
-# Set compiler optimization flags
-
-export {CFLAGS,CXXFLAGS}='-g -O3 -pipe -fno-plt -march=native'
-
+CC="gcc"
+CXX="g++"
+CFLAGS="-O2 -pipe -march=native"
+CXXFLAGS="$CFLAGS"
 PATH="/usr/lib/ccache:$PATH"
-PKG_CONFIG_PATH="/usr/local/lib64/pkgconfig:/usr/local/lib/pkgconfig:/usr/lib64/pkgconfig:/usr/lib/pkgconfig"
-PKG_CONFIG_PATH+=":/usr/local/lib/x86_64-linux-gnu/pkgconfig:/usr/lib/x86_64-linux-gnu/pkgconfig"
-export PATH PKG_CONFIG_PATH
+PKG_CONFIG_PATH="/usr/local/lib/pkgconfig:/usr/local/lib64/pkgconfig:/usr/local/share/pkgconfig:/usr/lib/pkgconfig:/usr/lib64/pkgconfig:/usr/share/pkgconfig"
+PKG_CONFIG_PATH+=":/usr/local/cuda/lib64/pkgconfig:/usr/local/cuda/lib/pkgconfig:/opt/cuda/lib64/pkgconfig:/opt/cuda/lib/pkgconfig"
+PKG_CONFIG_PATH+=":/usr/lib/x86_64-linux-gnu/pkgconfig:/usr/lib/i386-linux-gnu/pkgconfig:/usr/lib/arm-linux-gnueabihf/pkgconfig:/usr/lib/aarch64-linux-gnu/pkgconfig"
+export CC CXX CFLAGS CXXFLAGS PATH PKG_CONFIG_PATH
 
 # Create functions
